requirements.txt
streamlit
google-cloud-firestore
protobuf~=3.19.0
pandas
